+++ backend/src/core_telemetry/lib_config.py
@@ -22,6 +22,7 @@ from functools import cache
 from pathlib import Path
 
 from pydantic import (
+    BaseModel,
     Field,
     StringConstraints,
 )
@@ -36,6 +37,88 @@ JwtSecretStr = Annotated[str, StringConstraints(pattern='[0-9a-f]{32,160}')]
 MinimalStr = Annotated[str, StringConstraints(to_lower=True, min_length=3)]
 LowerCaseStr = Annotated[str, StringConstraints(to_lower=True)]
 PasswordOrKeyStr = Annotated[str, StringConstraints(min_length=8)]
+
+class TelemetrySettings(BaseModel):
+    """
+    Configuration group for OpenTelemetry settings.
+
+    This class encapsulates all OpenTelemetry-related configuration options,
+    including tracing, metrics, exporter endpoints, and environment details.
+    It is used as a nested model within LibrarySettings to organize telemetry
+    configuration in a structured way.
+    """
+    enable_opentelemetry: Annotated[
+        bool,
+        Field(
+            default=True,
+            validation_alias='ENABLE_OPENTELEMETRY',
+            description='Enable or disable OpenTelemetry instrumentation.'
+        )
+    ]
+    service_name: Annotated[
+        str,
+        Field(
+            default='answers-agent',
+            validation_alias='OTEL_SERVICE_NAME',
+            description='Service name for OpenTelemetry traces and metrics.'
+        )
+    ]
+    service_version: Annotated[
+        str,
+        Field(
+            default='0.1.0',
+            validation_alias='OTEL_SERVICE_VERSION',
+            description='Service version for OpenTelemetry.'
+        )
+    ]
+    environment: Annotated[
+        str,
+        Field(
+            default='development',
+            validation_alias='OTEL_ENVIRONMENT',
+            description='Deployment environment for OpenTelemetry (e.g., development, staging, production).'
+        )
+    ]
+    jaeger_endpoint: Annotated[
+        str,
+        Field(
+            default='http://jaeger:14268/api/traces',
+            validation_alias='OTEL_EXPORTER_JAEGER_ENDPOINT',
+            description='Jaeger endpoint for exporting OpenTelemetry traces.'
+        )
+    ]
+    prometheus_port: Annotated[
+        int,
+        Field(
+            default=8889,
+            validation_alias='OTEL_EXPORTER_PROMETHEUS_PORT',
+            description='Port for Prometheus metrics exporter.'
+        )
+    ]
+    enable_tracing: Annotated[
+        bool,
+        Field(
+            default=True,
+            validation_alias='OTEL_ENABLE_TRACING',
+            description='Enable or disable OpenTelemetry tracing.'
+        )
+    ]
+    enable_metrics: Annotated[
+        bool,
+        Field(
+            default=True,
+            validation_alias='OTEL_ENABLE_METRICS',
+            description='Enable or disable OpenTelemetry metrics.'
+        )
+    ]
+    trace_sample_rate: Annotated[
+        float,
+        Field(
+            default=1.0,
+            validation_alias='OTEL_TRACE_SAMPLE_RATE',
+            description='Sampling rate for OpenTelemetry traces (0.0 to 1.0).'
+        )
+    ]
 
 
 class LibrarySettings(BaseSettings):
@@ -83,18 +166,13 @@ class LibrarySettings(BaseSettings):
         )
 
     # OpenTelemetry configuration
-    enable_opentelemetry: bool = Field(default=True, validation_alias='ENABLE_OPENTELEMETRY')
-    otel_service_name: str = Field(default='answers-agent', validation_alias='OTEL_SERVICE_NAME')
-    otel_service_version: str = Field(default='0.1.0', validation_alias='OTEL_SERVICE_VERSION')
-    otel_environment: str = Field(default='development', validation_alias='OTEL_ENVIRONMENT')
-    otel_jaeger_endpoint: str = Field(
-        default='http://jaeger:14268/api/traces', validation_alias='OTEL_EXPORTER_JAEGER_ENDPOINT'
-    )
-    otel_prometheus_port: int = Field(default=8889, validation_alias='OTEL_EXPORTER_PROMETHEUS_PORT')
-    otel_enable_tracing: bool = Field(default=True, validation_alias='OTEL_ENABLE_TRACING')
-    otel_enable_metrics: bool = Field(default=True, validation_alias='OTEL_ENABLE_METRICS')
-    otel_trace_sample_rate: float = Field(default=1.0, validation_alias='OTEL_TRACE_SAMPLE_RATE')
-
+    otel: Annotated[
+        TelemetrySettings,
+        Field(
+            default_factory=TelemetrySettings,
+            description="OpenTelemetry configuration group."
+        )
+    ]
 
 @cache
 def get_lib_config():
